backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.34
asyncpg==0.29.0
pydantic==2.8.2
pydantic-settings==2.4.0
aiogram==3.13.1
PyJWT==2.9.0
python-dotenv==1.0.1